requirements.txt
jupyter
openai
langchain
langchain_community
langchain-ollama
anthropic
crewai
crewai_tools
python-dotenv
requests
tenacity
termcolor
pydantic
google-generativeai